Refactor: split ProducerStubFactory between public and private API
 * Moved private classes to .internal.testkit package
 * Refactor Fake Services to contain interface and Impl in a single file.
