+++ testkit/javadsl/src/test/java/com/lightbend/lagom/javadsl/testkit/services/FoxtrotDownstreamService.java
@@ -4,10 +4,15 @@
 package com.lightbend.lagom.javadsl.testkit.services;
 
 import akka.NotUsed;
+import akka.stream.Materializer;
 import com.lightbend.lagom.javadsl.api.Descriptor;
 import com.lightbend.lagom.javadsl.api.Service;
 import com.lightbend.lagom.javadsl.api.ServiceCall;
 import org.pcollections.PSequence;
+import org.pcollections.TreePVector;
+
+import javax.inject.Inject;
+import java.util.concurrent.CompletableFuture;
 
 import static com.lightbend.lagom.javadsl.api.Service.named;
 import static com.lightbend.lagom.javadsl.api.Service.namedCall;
@@ -21,9 +26,29 @@ public interface FoxtrotDownstreamService extends Service {
         return named("foxtrot")
                 .withCalls(
                         namedCall("retrieveMessagesF", this::retrieveMessagesF)
-                )
-                ;
-
+                );
     }
 
+    class Impl implements FoxtrotDownstreamService {
+
+        private volatile PSequence<ReceivedMessage> receivedMessages = TreePVector.empty();
+
+        @Inject
+        public Impl(AlphaUpstreamService alphaUpstreamService, Materializer materializer) {
+            alphaUpstreamService
+                    .messageTopic()
+                    .subscribe()
+                    .withGroupId("downstream-foxtrot")
+                    .atMostOnceSource()
+                    .runForeach(
+                            alphaEvent -> receivedMessages = receivedMessages.plus(new ReceivedMessage("A", alphaEvent.getCode())),
+                            materializer
+                    );
+        }
+
+        @Override
+        public ServiceCall<NotUsed, PSequence<ReceivedMessage>> retrieveMessagesF() {
+            return req -> CompletableFuture.completedFuture(receivedMessages);
+        }
+    }
 }
